requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anyio==4.4.0
attrs==23.2.0
botocore==1.34.131
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.95.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.2
jmespath==1.0.1
lazy_loader==0.4
lxml==5.2.2
multidict==6.0.5
networkx==3.3
numpy==2.0.0
openai==1.35.7
opencv-python==4.10.0.84
openpyxl==3.1.4
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pillow==10.3.0
pipdeptree==2.23.0
pycparser==2.22
pydantic==1.10.17
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
tifffile==2024.6.18
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.1
wrapt==1.16.0
yarl==1.9.4
